Fix waveform bugs
Reverting some changes (primarily initializing the uncertainty map from std::map<unsigned long, double>* ChannelKeyToTimingSigmaMap; to std::map<unsigned long, double> ChannelKeyToTimingSigmaMap; to match the LoadGeometry tool and how it saves the map to the store).

There was a problem in which the waveformsim could not grab the timing uncertainties from the map (they were blank for some reason - the size of the loaded map was 0). This resulted in the default uncertainty of 1.0ns being set for each phototube:

hit charge =  1.222399 p.e., hit time =  65.417686 for PMTID 462
[2]: No timing sigma for PMT 462, using default
PMTWaveformSim:
    hit charge =  0.680517 p.e., hit time =  47.844757 for PMTID 463
[2]: No timing sigma for PMT 463, using default
PMTWaveformSim:
    hit charge =  1.071871 p.e., hit time =  48.019901 for PMTID 463
[2]: No timing sigma for PMT 463, using default

It was perplexing and led to indeterminate behavior (re-making the tool with no changes led to completely different behavior). I assume the problem was due to the map being initialized NOT as a pointer, which is different from the way LoadGeometry handles it. 

Will continue testing to see if this setup is stable.